code/lizzie/python/mini_capstone/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
Pillow==9.3.0
requests==2.28.1
urllib3==1.26.13
